+++ src/completions/loopObjCompletion.ts
